benchmark for memory allocation time

+++ projects/xkrt/include/xkrt/logger/metric.h
@@ -5,7 +5,7 @@
 /*   Author: Romain PEREIRA <rpereira@anl.gov>                     .'* *.'    */
 /*                                                              __/_*_*(_     */
 /*   Created: 2025/02/18 15:08:36 by Romain PEREIRA            / _______ \    */
-/*   Updated: 2025/03/02 16:15:07 by Romain PEREIRA            \_)     (_/    */
+/*   Updated: 2025/03/04 21:58:21 by Romain PEREIRA            \_)     (_/    */
 /*                                                                            */
 /*   License: CeCILL 2.1                                                      */
 /*                                                                            */
@@ -21,7 +21,7 @@ extern "C" {
     uint64_t xkrt_get_nanotime(void);
 
     void xkrt_metric_byte(char * buffer, int bufsize, size_t nbytes);
-    void xkrt_metric_time(char * buffer, int bufsize, size_t ns);
+    void xkrt_metric_time(char * buffer, int bufsize, uint64_t ns);
     void xkrt_metric_bandwidth(char * buffer, int bufsize, size_t byte_per_sec);
 };
